DRY parseImports
